% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitargets_dea.R
\name{dea_accessors}
\alias{dea_accessors}
\alias{markers_table}
\alias{go_table}
\alias{go_plot_data}
\alias{go_genes_html}
\alias{volcano_plot}
\alias{pca_plot}
\alias{go_barplot}
\alias{go_cnetplot}
\title{Differential-expression and enrichment accessors for \link{scitargets_dea}}
\usage{
markers_table(x, ...)

go_table(x, ...)

go_plot_data(x, ...)

go_genes_html(x, ...)

volcano_plot(x, ...)

pca_plot(x, ...)

go_barplot(x, ...)

go_cnetplot(x, ...)
}
\arguments{
\item{x}{A \link{scitargets_dea} object.}

\item{...}{Passed to methods (e.g. \code{level}, \code{direction}, \code{ontology}).}
}
\value{
A data.frame, HTML string, or plot object depending on the generic.
}
\description{
Generics to extract tables and plots from a \link{scitargets_dea} object. Most
take optional \code{level} ("single_cell"/"pseudobulk"), \code{direction} ("up"/"down")
and \code{ontology} ("BP"/"CC"/"MF") arguments, defaulting to the first available.
}
