% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex_cell.R
\name{demultiplex_cell}
\alias{demultiplex_cell}
\title{Demultiplex cell}
\usage{
demultiplex_cell(obj = NULL, default_seed = 1234)
}
\arguments{
\item{obj}{A Seurat object with the cells HTO informations}

\item{default_seed}{Default seed used for reduction methods that need it. Not used if the function run inside a targets pipeline as in this case the steps seed (defined based on its name and the pipeline global seed) is used.}
}
\value{
A seurat object with metadata containing the demultiplexing result.
}
\description{
Demultiplex the cells using the cite-seq experiment HTO.
}
\examples{
\dontrun{
demultiplex_cell(path = ".", id = "my_project_id")
}
}
