% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_non_negative_cells.R
\name{extract_non_negative_cells}
\alias{extract_non_negative_cells}
\title{Extract the non-negative cells (singlets and doublets)}
\usage{
extract_non_negative_cells(
  obj = NULL,
  hto_ident = "HTO_classification.global",
  default_seed = 1234
)
}
\arguments{
\item{obj}{a Seurat object.}

\item{hto_ident}{The name of the column in obj metadata that contains the HTO demultiplexing results. Default to \code{HTO_classification.global}.}

\item{default_seed}{Default seed used for reduction methods that need it. Not used if the function run inside a targets pipeline as in this case the steps seed (defined based on its name and the pipeline global seed) is used.}
}
\value{
A seurat object filtered to contains only singlets and doublets
}
\description{
Filter the Seurat object to keep only the singlets and the doublets. Recompute UMAP and t-SNE before returning the object.
}
\examples{
\dontrun{
extract_non_negative_cells(obj = Seurat_obj)
}
}
