% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cell_and_run_reduction.R
\name{filter_cell_and_run_reduction}
\alias{filter_cell_and_run_reduction}
\title{Filter cell and run reduction}
\usage{
filter_cell_and_run_reduction(
  obj = NULL,
  min_nFeature_RNA = 200,
  max_nFeature_RNA = 4000,
  cutoff_percent_mt = 5,
  default_seed = 1234
)
}
\arguments{
\item{obj}{a Seurat object.}

\item{min_nFeature_RNA}{Minimum number of genes in the filtered cells.}

\item{max_nFeature_RNA}{Maximum number of genes in the filtered cells.}

\item{cutoff_percent_mt}{Percentage of count corresponding to mitochondrial genes used to filter the cells. Cells with that percentage or more will be removed.}

\item{default_seed}{Default seed used for reduction methods that need it. Not used if the function run inside a targets pipeline as in this case the steps seed (defined based on its name and the pipeline global seed) is used.}
}
\value{
A seurat object
}
\description{
Filter cells based on the number of genes they contains and the percentage of count represented by mitochondrial genes.
After the filtering, PCA, UMAP, and t-SNE are computed. The UMAP and t-SNE are based on the PCA (PC used are those whose standard deviation is superior to the median of all PC standard deviation).
}
\examples{
\dontrun{
filter_cell_and_run_reduction(obj = Seurat_obj)
}
}
