% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitargets_dea.R
\name{gsea_accessors}
\alias{gsea_accessors}
\alias{gsea_table}
\alias{gsea_barplot}
\title{GSEA accessors for \link{scitargets_dea}}
\usage{
gsea_table(x, ...)

gsea_barplot(x, ...)
}
\arguments{
\item{x}{A \link{scitargets_dea} object.}

\item{...}{Passed to methods (e.g. \code{level}, \code{collection}).}
}
\value{
\code{gsea_table} returns a data.frame; \code{gsea_barplot} returns a ggplot.
}
\description{
GSEA accessors for \link{scitargets_dea}
}
