% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdwgcna_report.R
\name{hdwgcna_me_dotplot}
\alias{hdwgcna_me_dotplot}
\title{Dot plot of module eigengenes across cell types}
\usage{
hdwgcna_me_dotplot(obj, group_by)
}
\arguments{
\item{obj}{An hdWGCNA Seurat object.}

\item{group_by}{Metadata column to group the dot plot by (e.g. the cell-type
clustering column).}
}
\value{
A \code{ggplot} object.
}
\description{
Adds the harmonised module eigengenes (hMEs) of the analysed group's modules
to the metadata and draws a \code{Seurat::DotPlot} across \code{group_by}.
}
