% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitargets_dea.R
\name{is_scitargets_dea}
\alias{is_scitargets_dea}
\title{Test whether an object is a \link{scitargets_dea}}
\usage{
is_scitargets_dea(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a \link{scitargets_dea}
}
