% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seurat_data_10X.R
\name{load_seurat_data_10X}
\alias{load_seurat_data_10X}
\title{Load seurat data}
\usage{
load_seurat_data_10X(path = ".", id = NULL)
}
\arguments{
\item{path}{relative or absolute path to the directory that contains another directory whose name correspond to \code{id} and contains cellranger output (barcodes.tsv.gz, features.tsv.gz, matrix.mtx.gz).}

\item{id}{used as the project name for the Seurat object.}
}
\value{
A Seurat object.
}
\description{
Load cite-seq data generated by 10X cellranger into Seurat
}
\examples{
\dontrun{
load_seurat_data(path = ".", id = "my_project_id")
}
}
