% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth_annot_pbmc.R
\name{azimuth_annot_pbmc}
\alias{azimuth_annot_pbmc}
\title{Azimuth annotations for PBMC}
\usage{
azimuth_annot_pbmc(obj, cluster_to_use = "clusters_0.5")
}
\arguments{
\item{obj}{a Seurat object with an SCT assay and a metadata field corresponding to cell clusters}

\item{cluster_to_use}{Name of the cells clusters}
}
\value{
A seurat object
}
\description{
Azimuth annotations for PBMC
}
\examples{
\dontrun{
azimuth_annot_pbmc(SeuratObj, "clusters_name_in_metadata")
}
}
