% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_singlets.R
\name{extract_singlets}
\alias{extract_singlets}
\title{Extract the singlets}
\usage{
extract_singlets(
  obj = NULL,
  dims_to_use = 1:15,
  feat_to_remove = NULL,
  default_seed = 1234
)
}
\arguments{
\item{obj}{obj a Seurat object.}

\item{dims_to_use}{Number of principal components to use when computing the UMAP and t-SNE. Default to 1:15. If a NULL or non-integer value is set, the number will be determined by the stdev of the PC.}

\item{feat_to_remove}{Names of features (genes) to remove. Can be used to remove genes located on gonosomes.}

\item{default_seed}{default_seed Default seed used for reduction methods that need it. Not used if the function run inside a targets pipeline as in this case the steps seed (defined based on its name and the pipeline global seed) is used.}
}
\value{
A Seurat object
}
\description{
Filter the Seurat object to keep only the singlets identified using the HTO and removes features (genes) if necessary before recomputing the UMAP and t-SNE and identify cluster for different resolution (0.2 to 1).
}
\examples{
\dontrun{
extract_singletss(obj = Seurat_obj)
}
}
