% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitargets-package.R
\docType{package}
\name{scitargets}
\alias{scitargets-package}
\alias{scitargets}
\title{scitargets: Single-cell CITE-seq analysis with targets}
\description{
This package provides target factories and utilities for analyzing
single-cell \href{https://en.wikipedia.org/wiki/CITE-Seq}{CITE-seq} data with the \pkg{Seurat} and \pkg{targets} packages.
}
\section{scitargets Functions}{
\subsection{Function used by the targets factories:}{
\enumerate{
\item \link{load_seurat_data_10X}: load CITE-seq data from 10X cellranger output
\item \link{filter_cell_and_run_reduction}: To be called on the loaded seurat data to filter cells based on their number of genes and percentage of mitochondrial counts.
\item \link{demultiplex_cell}: demultiplex the cells using the HTO assay
\item \link{extract_non_negative_cells}: filter the cells to remove the negative cells (keep both singlets and doublets).
\item \link{extract_singlets}: filter the cells to keep only the singlets. Also allow to remove some features (genes) by names before recomputing UMAP, t-SNE, and clustering the cells.
\item \link{azimuth_annot_pbmc}: Use azimuth to annotate celltypes using a PBMC reference
}
}

\subsection{Targets factories:}{

Targets factories are low-level functions used to generate an ensemble of targets steps. For more information about targets factories check the \href{https://wlandau.github.io/targetopia/contributing.html}{targetopia contribution section}:
\enumerate{
\item \link{tar_demultiplex_hto}: Used to create a pipeline using the functions 1 to 5 from the previous steps.
}
}

\subsection{Other targets related functions:}{
\enumerate{
\item \link{tar_visnetwork_enhanced}: Similar to \link[targets:tar_visnetwork]{targets::tar_visnetwork} but the graph was modified to easier to read.
}
}

\subsection{Exported functions:}{
\itemize{
\item magrittr pipes: \link[magrittr]{\%>\%}, \link[magrittr]{\%<>\%}
}
}
}

\author{
\strong{Maintainer}: Pierre Solomon \email{pierre.solomon@laposte.net}

}
