% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_demultiplex_hto.R
\name{tar_demultiplex_hto}
\alias{tar_demultiplex_hto}
\title{Targets factory to demultiplex CITE-seq}
\usage{
tar_demultiplex_hto(
  run_id = "ID_NAME",
  project_id = "PROJECT_ID",
  run_path = NA,
  min_genes_detected = 200,
  max_genes_detected = 4000,
  mt_percent_cutoff = 5,
  singlets_feat_to_remove = NULL,
  singlets_dim_to_use = 1:15,
  singlets_clusters_to_use = NULL,
  run_azimuth = F
)
}
\arguments{
\item{run_id}{Used to set the names of of the targets. Must be unique across all project}

\item{project_id}{Used as the project name for the Seurat object.}

\item{run_path}{Relative or absolute path to the directory that contains another directory whose name correspond to id and contains cellranger output (barcodes.tsv.gz, features.tsv.gz, matrix.mtx.gz).}

\item{min_genes_detected}{Minimal number of genes detected in a cell. Default to 200.}

\item{max_genes_detected}{Maximal number of genes detected in a cell. Default to 4000.}

\item{mt_percent_cutoff}{Cutoff for the percentage of mitochondrial counts a cell can have.}

\item{singlets_feat_to_remove}{If not NULL, a character variable indicating a targets step that contains the name of features (genes) to remove from the singlets.}

\item{singlets_dim_to_use}{PCA dimensions to use when computing UMAP and t-SNE (only used after extracting the singlets cells).}

\item{singlets_clusters_to_use}{Name of cell cluster to use when identifying markers or running azimuth celltype annotation}

\item{run_azimuth}{If true, run azimuth using the (human) pbmc-ref of SeuratData}
}
\value{
A list describing pipeline steps that can be interpreted by targets
}
\description{
This function is used to create the pipeline steps necessary to:
\itemize{
\item load the CITE-seq data
\item filter the assay based on the number of detected genes and percentage of mitochondrial counts
\item demultiplex the HTO
\item extract non-negative (singlets + doublets) cells
\item extract singlets cells.
\item remove non-desired features (gonosomes genes) from singlets
\item realize a celltype annotation on the singlets using azimuth
}
}
\details{
Note that if we remove features, azimuth will be run in the version without those features.

Dimension reduction (PCA, UMAP, t-SNE) are computed after each modification of the Seurat object that modify the count matrix.
}
\examples{
\dontrun{
# Check the "get-started" and "targets-factories" vignettes for examples
}
}
