% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_visnetwork_enhanced.R
\name{tar_visnetwork_enhanced}
\alias{tar_visnetwork_enhanced}
\title{Enhanced visNetwork dependency graph}
\usage{
tar_visnetwork_enhanced(
  color_stem = "#75AADB",
  color_standalone = "#F4A261",
  color_function = "#9CCC65",
  color_pattern = "#E5989B",
  redirect_stem_to_function = TRUE
)
}
\arguments{
\item{color_stem}{Color for stems (targets steps) that depends of other steps}

\item{color_standalone}{Color for stems that don't depends of other steps (input files)}

\item{color_function}{Color for the functions used by the pipeline}

\item{color_pattern}{Color for the steps that use dynamic/static branching}

\item{redirect_stem_to_function}{If set to true and a stem depends of both a function and another stem, then we redirect the edge to the function.}
}
\value{
A HTML widget with a visNetwork graph representing the project targets pipeline.
}
\description{
Enhanced version of the \code{targets::tar_visnetwork} function.
}
\details{
Visualize the pipeline dependency graph with a visNetwork HTML widget.
}
\examples{
\dontrun{
tar_visnetwork_enhanced()
}
}
